Library

Documenting the user interface.

Wall stress model

Walrus.WallStressModel — Module

WallStressModel

A wall stress model for LES simulation with default parameters similar to that proposed in Schumann (1975), Hartel (1996), Piomelli et al. (1989), and Taylor and Sarkar (2007).

Walrus.WallStressModel.WallStress — Type

WallStress(; von_Karman_constant = 0.4,
             kinematic_viscosity = 1e-6,
             B = 5.2)

Returns a wall stress model for LES simulation with default parameters similar to that proposed in Schumann (1975), Hartel (1996), Piomelli et al. (1989), and Taylor and Sarkar (2007).

Keyword Arguments

von_Karman_constant: the von Karman wall stress constant

kinematic_viscosity: kinematic viscosity of the water above the wall

B: wall stress constant

Example

julia> using Walrus: WallStress

julia> using Oceananigans

julia> wall_stress = WallStress()
(::WallStress{Float64}) (generic function with 1 method)
julia> boundary_conditions = (u = FieldBoundaryConditions(bottom = FluxBoundaryCondition(wall_stress, discrete_form = true, parameters = :x)),
                              v = FieldBoundaryConditions(bottom = FluxBoundaryCondition(wall_stress, discrete_form = true, parameters = :y)))
(u = Oceananigans.FieldBoundaryConditions, with boundary conditions
├── west: DefaultBoundaryCondition (FluxBoundaryCondition: Nothing)
├── east: DefaultBoundaryCondition (FluxBoundaryCondition: Nothing)
├── south: DefaultBoundaryCondition (FluxBoundaryCondition: Nothing)
├── north: DefaultBoundaryCondition (FluxBoundaryCondition: Nothing)
├── bottom: FluxBoundaryCondition: DiscreteBoundaryFunction (::WallStress{Float64}) with parameters Symbol
├── top: DefaultBoundaryCondition (FluxBoundaryCondition: Nothing)
└── immersed: DefaultBoundaryCondition (FluxBoundaryCondition: Nothing), v = Oceananigans.FieldBoundaryConditions, with boundary conditions
├── west: DefaultBoundaryCondition (FluxBoundaryCondition: Nothing)
├── east: DefaultBoundaryCondition (FluxBoundaryCondition: Nothing)
├── south: DefaultBoundaryCondition (FluxBoundaryCondition: Nothing)
├── north: DefaultBoundaryCondition (FluxBoundaryCondition: Nothing)
├── bottom: FluxBoundaryCondition: DiscreteBoundaryFunction (::WallStress{Float64}) with parameters Symbol
├── top: DefaultBoundaryCondition (FluxBoundaryCondition: Nothing)
└── immersed: DefaultBoundaryCondition (FluxBoundaryCondition: Nothing))

Walrus.WallStressModel.WallStressBoundaryConditions — Method

WallStressBoundaryConditions(; von_Karman_constant = 0.4,
                               kinematic_viscosity = 1e-6,
                               B = 5.2)

Convenience constructor to setup WallStress boundary conditions.

Keyword Arguments

von_Karman_constant: the von Karman wall stress constant

kinematic_viscosity: kinematic viscosity of the water above the wall

B: wall stress constant

Example

julia> using Walrus: WallStressBoundaryConditions

julia> using Oceananigans

julia> stress_boundary_conditions = WallStressBoundaryConditions()
(u = FluxBoundaryCondition: DiscreteBoundaryFunction (::WallStress{Float64}) with parameters Symbol, v = FluxBoundaryCondition: DiscreteBoundaryFunction (::WallStress{Float64}) with parameters Symbol)
julia> boundary_conditions = (u = FieldBoundaryConditions(bottom = stress_boundary_conditions.u),
                              v = FieldBoundaryConditions(bottom = stress_boundary_conditions.v))
(u = Oceananigans.FieldBoundaryConditions, with boundary conditions
├── west: DefaultBoundaryCondition (FluxBoundaryCondition: Nothing)
├── east: DefaultBoundaryCondition (FluxBoundaryCondition: Nothing)
├── south: DefaultBoundaryCondition (FluxBoundaryCondition: Nothing)
├── north: DefaultBoundaryCondition (FluxBoundaryCondition: Nothing)
├── bottom: FluxBoundaryCondition: DiscreteBoundaryFunction (::WallStress{Float64}) with parameters Symbol
├── top: DefaultBoundaryCondition (FluxBoundaryCondition: Nothing)
└── immersed: DefaultBoundaryCondition (FluxBoundaryCondition: Nothing), v = Oceananigans.FieldBoundaryConditions, with boundary conditions
├── west: DefaultBoundaryCondition (FluxBoundaryCondition: Nothing)
├── east: DefaultBoundaryCondition (FluxBoundaryCondition: Nothing)
├── south: DefaultBoundaryCondition (FluxBoundaryCondition: Nothing)
├── north: DefaultBoundaryCondition (FluxBoundaryCondition: Nothing)
├── bottom: FluxBoundaryCondition: DiscreteBoundaryFunction (::WallStress{Float64}) with parameters Symbol
├── top: DefaultBoundaryCondition (FluxBoundaryCondition: Nothing)
└── immersed: DefaultBoundaryCondition (FluxBoundaryCondition: Nothing))

Radiative transfer models

Walrus.RadiativeTransfer — Module

RadiativeTransfer

Includes models for ratiative transfer through water which can induce body heating

Walrus.RadiativeTransfer.BodyHeating — Type

BodyHeating

A model for single band light attenuation which heats the water in the form: $\frac{\partial I}{\partial z} = -\alpha I$ where $I$ is the radiation intensity and $\alpha$ is the attenuation coefficient. This heats the water like $\frac{\partial T}{\partial t} = \frac{\alpha I}{c^p\rho}$ where $c^p$ is the specific heat capacity and $\rho$ is the water density.

Walrus.RadiativeTransfer.BodyHeating — Method

BodyHeating(; surface_flux,
              grid = nothing,
              field = nothing,
              water_attenuation_coefficient = 1.8,
              water_heat_capacity = 3991.0, # J K⁻¹ kg⁻¹
              water_density = 1026.0) # kg m⁻³

Creates a model in which a surface_flux (W / m²) is attenuated by and heats the water. This interacts with Oceananigans through the biogeochemistry but can also be used as a modifier in OceanBioME's Biogeochemistry.

Keyword Arguments

surface_flux (required): a function returning the surface radiaiton flux in the form surface_flux(x, y, t)

grid: the model grid required to setup the radiaiton field (reccommended)

field: alternativly a pre defined radiation field

water_attenuation_coefficient: the radiation attenuation coefficient of the water

water_heat_capacity: the specific heat capacity of the water

water_density: density of the water

Example

julia> using Walrus: BodyHeating

julia> using Oceananigans

julia> grid = RectilinearGrid(size = (128, 128, 128), extent = (1000, 1000, 1000))
128×128×128 RectilinearGrid{Float64, Oceananigans.Grids.Periodic, Oceananigans.Grids.Periodic, Oceananigans.Grids.Bounded} on Oceananigans.Architectures.CPU with 3×3×3 halo
├── Periodic x ∈ [0.0, 1000.0)  regularly spaced with Δx=7.8125
├── Periodic y ∈ [0.0, 1000.0)  regularly spaced with Δy=7.8125
└── Bounded  z ∈ [-1000.0, 0.0] regularly spaced with Δz=7.8125
julia> body_heating = BodyHeating(; surface_flux = (x, y, t) -> 100, grid)
┌ Warning: This radiative heating model is untested
└ @ Walrus.RadiativeTransfer ~/Documents/Projects/Walrus.jl/src/radiative_transfer/body_heating.jl:88
Single band light attenuation and body heating model with: 
- Attenuation coefficient: 1.8 (1 / m)
- Heat capacity: 3991.0 (J / K / kg)

julia> model = NonhydrostaticModel(; grid, biogeochemistry = body_heating, tracers = :T)
NonhydrostaticModel{CPU, RectilinearGrid}(time = 0 seconds, iteration = 0)
├── grid: 128×128×128 RectilinearGrid{Float64, Oceananigans.Grids.Periodic, Oceananigans.Grids.Periodic, Oceananigans.Grids.Bounded} on Oceananigans.Architectures.CPU with 3×3×3 halo
├── timestepper: QuasiAdamsBashforth2TimeStepper
├── tracers: T
├── closure: Nothing
├── buoyancy: Nothing
└── coriolis: Nothing

Tidal forcing models

Walrus.TidalForcings — Module

TidalForcing

Provides quick setup of tidal forcing

Walrus.TidalForcings.Tide — Type

Tide(; x_amplitude, 
       y_amplitude, 
       period = 12.3782216453hours,
       nodal_time = 0., 
       x_lag = 0., 
       y_lag = 0.,
       coriolis = nothing)

Sets up a model of tidal forcing with default parameters of an M2 tide.

Keyword Arguments

x_amplitude: the tidal amplitude in the x direction

y_amplitude: the tidal amplitude in the x direction

period: the tidal period (defaults to that of an M2 tide)

nodal_time: the time at which peak flow occurs

x_lag: the phase lag for the tidal component in the x direction

y_lag: the phase lag for the tidal component in the y direction

coriolis: a model for the coriolis parameter

Example

julia> using Walrus: Tide

julia> using Oceananigans

julia> tide = Tide(x_amplitude = 0.1, y_amplitude = 0.)
(::Tide{Float64, Nothing}) (generic function with 2 methods)
julia> forcing = (u = Forcing(tide, parameters = Val(:x), discrete_form = true),
                  v = Forcing(tide, parameters = Val(:y), discrete_form = true))
(u = DiscreteForcing{Val{:x}}
├── func: (::Tide{Float64, Nothing}) (generic function with 2 methods)
└── parameters: Val{:x}(), v = DiscreteForcing{Val{:y}}
├── func: (::Tide{Float64, Nothing}) (generic function with 2 methods)
└── parameters: Val{:y}())

Walrus.TidalForcings.TidalForcing — Method

TidalForcing(; x_amplitude,
               y_amplitude,
               period = 12.3782216453hours,
               nodal_time = 0.,
               x_lag = 0.,
               y_lag = 0.,
               coriolis = nothing)

A convenience constructor for Tide which returns the forcings pre wrapped.

Keyword Arguments

x_amplitude: the tidal amplitude in the x direction

y_amplitude: the tidal amplitude in the x direction

period: the tidal period (defaults to that of an M2 tide)

nodal_time: the time at which peak flow occurs

x_lag: the phase lag for the tidal component in the x direction

y_lag: the phase lag for the tidal component in the y direction

coriolis: a model for the coriolis parameter

Example

julia> using Walrus: TidalForcing

julia> tidal_forcing = TidalForcing(x_amplitude = 0.1, y_amplitude = 0.)
(u = DiscreteForcing{Val{:x}}
├── func: (::Tide{Float64, Nothing}) (generic function with 2 methods)
└── parameters: Val{:x}(), v = DiscreteForcing{Val{:y}}
├── func: (::Tide{Float64, Nothing}) (generic function with 2 methods)
└── parameters: Val{:y}())